[Bronze IV] Title: Do Not Touch Anything, Time: 64 ms, Memory: 5544 KB -BaekjoonHub

+++ 백준/Bronze/13136. Do Not Touch Anything/Do Not Touch Anything.cs
@@ -3,9 +3,9 @@ using System.Numerics;
 string input = Console.ReadLine();
 string[] splitInput = input.Split(' ');
 
-BigInteger r = BigInteger.Parse(splitInput[0]);
-BigInteger c = BigInteger.Parse(splitInput[1]);
-BigInteger n = BigInteger.Parse(splitInput[2]);
+int r = int.Parse(splitInput[0]);
+int c = int.Parse(splitInput[1]);
+int n = int.Parse(splitInput[2]);
 
 BigInteger row = r/n + (r%n == 0 ? 0 : 1);
 BigInteger col = c/n + (c%n == 0 ? 0 : 1);
